skyboxes and lighting
added some SFX for the player and added skyboxes in our folder. We can use them to make our game feel more alive. There is also some extra code singleton code for a audio manager. Use it as you like. Also tweaked a small bit of the lighting and shadows and stuff