functional rock generation
I feel like I am a legend at programming for doing this. I have looked extensively into figuring out how perlin noise works, and altough I feel like this isnt the perfect application for it, im really proud of making it work. After all the issues with unity before even starting to work on this, it came out better then I expected. I will prolly be tweaking this later on but this, for now, is a really easy way to do terrain generation.