fixed danis stupid skeletons
Dani had alot of issues with skeletons that were done. After testing the skeletons, there were still some issues to be had. Mostly the animations and the movement that was not working. After going through the code I left some comments with an explaination of what I changed or why I changed it. The skeleton is now fully functional and ready to be used in our demo