headache of the divine pillar
this divine pillar was my nemesis of life. This pillar was amazingly made on its own, but it did not work with other code AT ALL. I had to rework alot of the code to make it work on the current system. This was a horrible task as the update part of the methods had to be transformed to coroutines that will be activated on a coroutine manager. Otherwise these would not start as the main object was not active. To get this done was not that hard, but the fact I was not aware that the changed variables would stay like that even between playtime absolutely fucked me over. I have never wanted to kill myself ever this much to figure out how this works as even chatGPT was barely able to help me. i am so over this part and am happy I am done with this. zico out kms