item details
after working hard on the fixups for enemies and attack, I have finally realized the way to do Item details. I made use of mostly a method inside the Item classes that returns a dictionary with said details about the Items. The fields that are required can be added manually.