fixing up spell damage and enemyskeleton
To finish up most of the game I had to make the spells work on players and enemies alike. To do so I had to make a few changes:
I started off making a new script for every Entity in the game with health resistance and a damage method. I then put this on both enemies and players. This will for future endeavors double function as stats. We can alter stats an anything else from this base class and inherit it for example: enemies, players, ncps.
With some spells it requires an ontriggerenter. which needed an rigidbody to work. This was not yet on the skeleton and had to be added. I also set up an rebaked the navmesh for our world.