FloorPos edit, Jump and spawner
As i was finishing up tasks for the preview, I found out that the FloorPos method was not working properly. I was using layermask instead of layer.getmask. After fixing that, I fixed up the jumpCheck script that should be in the update method to work properly. After doing so and realising the spawner used it aswell. I had to fix that up because the enemies started spawning in the sky. -Z