EnemyEntity
I thought that our current situation with the entity script would encapsulate too much. As such I decided to split up the scripts a bit more and make them more specific to players and enemies. Then we have a script left for non-enemy Entities that can be used for most items in the game.